+++ apps/docs/sidebars.ts
@@ -103,17 +103,6 @@ const sidebars: SidebarsConfig = {
         },
         {
           type: 'category',
-          label: 'Numbers',
-          items: [
-            {
-              type: 'doc',
-              id: 'components/numbers/RollingNumber/rollingNumber',
-              label: 'RollingNumber',
-            },
-          ],
-        },
-        {
-          type: 'category',
           label: 'Inputs',
           items: [
             {
@@ -648,6 +637,17 @@ const sidebars: SidebarsConfig = {
         },
         {
           type: 'category',
+          label: 'Numbers',
+          items: [
+            {
+              type: 'doc',
+              id: 'components/numbers/RollingNumber/rollingNumber',
+              label: 'RollingNumber',
+            },
+          ],
+        },
+        {
+          type: 'category',
           label: 'Animation',
           items: [
             {
@@ -820,7 +820,7 @@ const sidebars: SidebarsConfig = {
       items: [
         {
           type: 'link',
-          href: 'https://github.com/coinbase/cds-staging/blob/master/packages/common/CHANGELOG.md',
+          href: 'https://github.com/coinbase/cds/blob/master/packages/common/CHANGELOG.md',
           label: '@coinbase/cds-common',
           customProps: {
             kbar: {
@@ -830,7 +830,7 @@ const sidebars: SidebarsConfig = {
         },
         {
           type: 'link',
-          href: 'https://github.com/coinbase/cds-staging/blob/master/packages/mobile/CHANGELOG.md',
+          href: 'https://github.com/coinbase/cds/blob/master/packages/mobile/CHANGELOG.md',
           label: '@coinbase/cds-mobile',
           customProps: {
             kbar: {
@@ -840,7 +840,7 @@ const sidebars: SidebarsConfig = {
         },
         {
           type: 'link',
-          href: 'https://github.com/coinbase/cds-staging/blob/master/packages/web/CHANGELOG.md',
+          href: 'https://github.com/coinbase/cds/blob/master/packages/web/CHANGELOG.md',
           label: '@coinbase/cds-web',
           customProps: {
             kbar: {
@@ -850,7 +850,7 @@ const sidebars: SidebarsConfig = {
         },
         {
           type: 'link',
-          href: 'https://github.com/coinbase/cds-staging/blob/master/packages/icons/CHANGELOG.md',
+          href: 'https://github.com/coinbase/cds/blob/master/packages/icons/CHANGELOG.md',
           label: '@coinbase/cds-icons',
           customProps: {
             kbar: {
@@ -860,7 +860,7 @@ const sidebars: SidebarsConfig = {
         },
         {
           type: 'link',
-          href: 'https://github.com/coinbase/cds-staging/blob/master/packages/illustrations/CHANGELOG.md',
+          href: 'https://github.com/coinbase/cds/blob/master/packages/illustrations/CHANGELOG.md',
           label: '@coinbase/cds-illustrations',
           customProps: {
             kbar: {
@@ -870,7 +870,7 @@ const sidebars: SidebarsConfig = {
         },
         {
           type: 'link',
-          href: 'https://github.com/coinbase/cds-staging/blob/master/packages/mobile-visualization/CHANGELOG.md',
+          href: 'https://github.com/coinbase/cds/blob/master/packages/mobile-visualization/CHANGELOG.md',
           label: '@coinbase/cds-mobile-visualization',
           customProps: {
             kbar: {
@@ -880,7 +880,7 @@ const sidebars: SidebarsConfig = {
         },
         {
           type: 'link',
-          href: 'https://github.com/coinbase/cds-staging/blob/master/packages/web-visualization/CHANGELOG.md',
+          href: 'https://github.com/coinbase/cds/blob/master/packages/web-visualization/CHANGELOG.md',
           label: '@coinbase/cds-web-visualization',
           customProps: {
             kbar: {