+++ packages/web/src/chips/TabbedChips.tsx
@@ -20,13 +20,14 @@ const scrollContainerCss = css`
 
 const TabComponent = <T extends string = string>({ label = '', id, ...tabProps }: TabValue<T>) => {
   const { activeTab, updateActiveTab } = useTabsContext();
-  const isActive = activeTab?.id === id;
+  const isActive = useMemo(() => activeTab?.id === id, [activeTab, id]);
   const handleClick = useCallback(() => updateActiveTab(id), [id, updateActiveTab]);
   return (
     <Chip
-      aria-checked={isActive}
+      aria-selected={isActive}
       inverted={isActive}
       onClick={handleClick}
+      role="tab"
       width="max-content"
       {...tabProps}
     >
@@ -59,7 +60,7 @@ const TabbedChipsComponent = memo(
       testID,
       background = 'bg',
       gap = 1,
-      role = 'radiogroup',
+      role,
       previousArrowAccessibilityLabel = 'Previous',
       nextArrowAccessibilityLabel = 'Next',
       width = '100%',